requirements.txt
pytest
aiogram
python-dotenv
aiosqlite
dataclasses-json
flask
